fix: CompanyDetail rename
