+++ src/components/company/CompanyDetail.vue
